Update R system

+++ Data Analysis/Class/DiscreteGrid.cs
@@ -3,7 +3,7 @@
     public class DiscreteGrid
     {
         public double number { get; set; }
-        public int frequency { get; set; }
+        public double frequency { get; set; }
         public int accumulatedFrequency { get; set; }
     }
 }
